Json generator fixed on random factory
