+++ pom.xml
@@ -92,6 +92,9 @@
         <maven-deploy-plugin.version>2.8.2</maven-deploy-plugin.version>
         <checkstyle.version>2.17</checkstyle.version>
         <spotbugs.version>4.1.3</spotbugs.version>
+        <exec-maven-plugin.version>3.0.0</exec-maven-plugin.version>
+
+        <gson.cache.folder>${main.basedir}/modules/Fading Suns Revised Edition/json</gson.cache.folder>
     </properties>
 
     <build>
