+++ think-machine-random/src/main/java/com/softwaremagico/tm/cache/ModuleLoaderEnforcer.java
@@ -1,4 +1,4 @@
-package com.softwaremagico.tm.random.cache;
+package com.softwaremagico.tm.cache;
 
 /*-
  * #%L
@@ -8,17 +8,17 @@ package com.softwaremagico.tm.random.cache;
  * %%
  * This software is designed by Jorge Hortelano Otero. Jorge Hortelano Otero
  * <softwaremagico@gmail.com> Valencia (Spain).
- *  
+ *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation; either version 2 of the License, or (at your option) any later
  * version.
- *  
+ *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
  * details.
- *  
+ *
  * You should have received a copy of the GNU General Public License along with
  * this program; If not, see <http://www.gnu.org/licenses/gpl-3.0.html>.
  * #L%
@@ -66,8 +66,9 @@ public class ModuleLoaderEnforcer {
 
 
         final long duration = (System.nanoTime() - startTime);
-        MachineLog.info(com.softwaremagico.tm.file.modules.ModuleLoaderEnforcer.class.getName(), "All factories loaded! Total {} elements loaded in {} milliseconds.",
+        MachineLog.info(com.softwaremagico.tm.file.modules.ModuleLoaderEnforcer.class.getName(),
+                "All factories loaded! Total {} elements loaded in {} milliseconds.",
                 loadedElements.get(), duration / 1000000);
-        return   loadedElements.get();
+        return loadedElements.get();
     }
 }
